js-fullstack/cycle-stage-model-view: front: Reenabled router for the App.
Since I'm no longe using the MVI helper (because it was not working with my past
implementation of the router) I decided to roll-back to the cyclic-router implementation.
Discovered in the way, how to actually use it (lol) turns out the routing is made via
javascript only and it cannot be replicated otherwise (I guess thats obvious now since
the app is not isomorphic). There's still need to find a way to actually persist the
state, but that's for another session.